Added ISMIR
ISMIR sees submissions with heavy application of ML to music in the form of both audio signals and symbolic representations. It's an A1 conference. IMHO it falls somewhere between subs ML, SP, and NLP, but I've given it the sub SP (which betrays the swathes of NLP like applications to symbolic representations of music :flushed:).